Delete unused hkdf.rs: migrate tests to crypto::tls13
